@@ -10,6 +10,7 @@
 		"check": "tsc --noEmit",
 		"format": "prettier --write \"**/*.{cjs,graphql,js,ts,json}\"",
 		"format:check": "prettier --check .",
+		"lint": "pnpm eslint .",
 		"prisma:generate": "prisma generate",
 		"test": "uvu -r tsm tests"
 	},
@@ -17,27 +18,30 @@
 	"author": "Rodney Lab <ask@rodneylab.com> (@rodneyLab)",
 	"license": "BSD-3-Clause",
 	"devDependencies": {
-		"@commitlint/config-conventional": "^19.2.2",
-		"@types/node": "^20.12.8",
-		"@typescript-eslint/eslint-plugin": "^7.8.0",
-		"@typescript-eslint/parser": "^7.8.0",
-		"commitlint": "^19.3.0",
-		"dotenv": "^16.4.5",
-		"eslint": "^9.1.1",
-		"eslint-plugin-import": "^2.29.1",
-		"husky": "^9.0.11",
-		"prettier": "^3.2.5",
-		"prisma": "^5.13.0",
+		"@commitlint/config-conventional": "^19.7.1",
+		"@types/node": "^22.13.0",
+		"@typescript-eslint/eslint-plugin": "^8.23.0",
+		"@typescript-eslint/parser": "^7.18.0",
+		"commitlint": "^19.7.1",
+		"dotenv": "^16.4.7",
+		"eslint": "^9.19.0",
+		"eslint-config-prettier": "^10.0.1",
+		"eslint-plugin-import": "^2.31.0",
+		"globals": "^15.14.0",
+		"husky": "^9.1.7",
+		"prettier": "^3.4.2",
+		"prisma": "^6.3.0",
 		"ts-node-dev": "^2.0.0",
 		"tsm": "^2.3.0",
-		"typescript": "^5.4.5",
+		"typescript": "^5.7.3",
+		"typescript-eslint": "^8.23.0",
 		"uvu": "^0.5.6"
 	},
 	"dependencies": {
-		"@prisma/client": "^5.13.0",
-		"fastify": "^4.26.2",
-		"graphql": "^16.8.1",
-		"mercurius": "^14.1.0",
+		"@prisma/client": "^6.3.0",
+		"fastify": "^5.2.1",
+		"graphql": "^16.10.0",
+		"mercurius": "^16.0.1",
 		"nexus": "^1.3.0"
 	},
 	"peerDependencies": {
