@@ -11,10 +11,10 @@ declare global {
 }
 
 if (process.env.NODE_ENV === 'production') {
-	db = new PrismaClient();
+	db = new PrismaClient({ log: ['query'] });
 } else {
 	if (!global.__db) {
-		global.__db = new PrismaClient();
+		global.__db = new PrismaClient({ log: ['query'] });
 	}
 	db = global.__db;
 }
